pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.19.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit-folium>=0.24.0",
    "streamlit>=1.42.0",
    "openai>=1.63.0",
]
